Move Keychain implementation to Keychain.swift
